Refactor package name and add LoggedShellAsserter implementation.
